+++ internal/paw/item_ssh_key.go
@@ -11,6 +11,8 @@ type SSHKey struct {
 	*Metadata `json:"metadata,omitempty"`
 	*Note     `json:"note,omitempty"`
 
+	AddToAgent  bool   `json:"add_to_agent,omitempty"`
+	Comment     string `json:"comment,omitempty"`
 	Fingerprint string `json:"fingerprint,omitempty"`
 	PrivateKey  string `json:"private_key,omitempty"`
 	PublicKey   string `json:"public_key,omitempty"`